move GetTeamByIDAndUserIDAuth to new file
